+++ pub/file/lib/src/interface/file_system.dart
@@ -10,13 +10,29 @@ abstract class FileSystem {
   const FileSystem();
 
   /// Returns a reference to a [Directory] at [path].
-  Directory directory(String path);
+  ///
+  /// [path] can be either a [`String`], a [`Uri`], or a [`FileSystemEntity`].
+  Directory directory(path);
 
   /// Returns a reference to a [File] at [path].
-  File file(String path);
+  ///
+  /// [path] can be either a [`String`], a [`Uri`], or a [`FileSystemEntity`].
+  File file(path);
 
   /// Returns a reference to a [Link] at [path].
-  Link link(String path);
+  ///
+  /// [path] can be either a [`String`], a [`Uri`], or a [`FileSystemEntity`].
+  Link link(path);
+
+  /// Gets the path separator used by this file system to separate components
+  /// in file paths.
+  String get pathSeparator;
+
+  /// Gets the system temp directory.
+  ///
+  /// It is left to file system implementations to decide how to define the
+  /// "system temp directory".
+  Directory get systemTempDirectory;
 
   /// Creates a directory object pointing to the current working directory.
   Directory get currentDirectory;