+++ pub/file/lib/src/io/shim_dart_io.dart
@@ -2,9 +2,13 @@ import 'dart:async';
 
 import 'dart:io' as io;
 
-io.Directory newDirectory(String path) => new io.Directory(path);
-io.File newFile(String path) => new io.File(path);
-io.Link newLink(String path) => new io.Link(path);
+import 'package:file/src/common.dart' as common;
+
+io.Directory newDirectory(path) => new io.Directory(common.getPath(path));
+io.File newFile(path) => new io.File(common.getPath(path));
+io.Link newLink(path) => new io.Link(common.getPath(path));
+String get pathSeparator => io.Platform.pathSeparator;
+io.Directory systemTemp() => io.Directory.systemTemp;
 io.Directory get currentDirectory => io.Directory.current;
 set currentDirectory(dynamic path) => io.Directory.current = path;
 Future<io.FileStat> stat(String path) => io.FileStat.stat(path);